requirements.txt
huggingface_hub
pysimplegui
rapidfuzz
python-dotenv
pdfplumber
sqlite-utils
matplotlib